+++ NEWS.md
@@ -1,5 +1,9 @@
 # rvest (development version)
 
+* `html_table()` correctly handles tables with cells that contain blank values 
+  for `rowspan` and/or `colspan`, so that e.g. `<td rowspan="">` is parsed as 
+  `<td rowspan=1>` (@epiben, #323).
+
 # rvest 1.0.0
 (to be released as rvest 1.0.0)